+++ topics/main.go
@@ -1,0 +1,10 @@
+package topics
+
+func init() {
+	Register("mem", NewMemProvider())
+	p, err := NewPudgeProvider(DefaultDbPath, DefaultSyncIntv)
+	if err != nil {
+		panic("cannot register pudge provider")
+	}
+	Register("pudge", p)
+}
